Feature/#9 GitHub actions

+++ src/Controller/Adminhtml/Message/MassDelete.php
@@ -9,7 +9,8 @@ use Magento\Backend\App\Action\Context;
 use Magento\Framework\Controller\Result\Redirect;
 use Magento\Framework\Controller\Result\RedirectFactory;
 use Magento\Ui\Component\MassAction\Filter;
-use RunAsRoot\MessageQueueRetry\Model\ResourceModel\Message\CollectionFactory;
+use RunAsRoot\MessageQueueRetry\Model\Message;
+use RunAsRoot\MessageQueueRetry\Model\ResourceModel\Message\MessageCollectionFactory;
 use RunAsRoot\MessageQueueRetry\Repository\MessageRepository;
 
 class MassDelete extends Action
@@ -20,7 +21,7 @@ class MassDelete extends Action
         Context $context,
         private MessageRepository $messageRepository,
         private RedirectFactory $redirectFactory,
-        private CollectionFactory $collectionFactory,
+        private MessageCollectionFactory $collectionFactory,
         private Filter $filter
     ) {
         parent::__construct($context);
@@ -34,13 +35,17 @@ class MassDelete extends Action
             $collection = $this->filter->getCollection($this->collectionFactory->create());
 
             foreach ($collection->getItems() as $message) {
+                if (!$message instanceof Message) {
+                    continue;
+                }
+
                 $this->messageRepository->delete($message);
             }
 
-            $this->messageManager->addSuccessMessage(__('The messages have been successfully deleted'));
+            $this->messageManager->addSuccessMessage(__('The messages have been successfully deleted')->render());
         } catch (\Exception $e) {
             $this->messageManager->addErrorMessage(
-                __('An error occurred while trying to delete the messages: %1', $e->getMessage())
+                __('An error occurred while trying to delete the messages: %1', $e->getMessage())->render()
             );
         }
